requirements.txt
Django==4.1.4
djangorestframework==3.14.0
drf-spectacular>=0.26.5
psycopg2-binary==2.9.3
python-dotenv==1.0.1
pika==1.3.2
redis==5.0.3
